Eagerly pull images in parallel
