+++ Cargo.toml
@@ -9,6 +9,7 @@ ctrlc = "3"
 env_logger = "0.6"
 hex = "0.3"
 log = "0.4"
+rayon = "1"
 scopeguard = "1"
 serde_yaml = "0.8"
 sha2 = "0.8"
